fix controller searching byc customerId

+++ customer-service/src/main/java/com/paint/customerservice/repository/impl/CustomerRepositoryCustom.java
@@ -1,5 +1,6 @@
 package com.paint.customerservice.repository.impl;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import com.paint.customerservice.model.Customer;
@@ -7,5 +8,5 @@ import com.paint.customerservice.model.Customer;
 public interface CustomerRepositoryCustom{
     List<Customer> findActiveCustomer();
     Customer findActiveCustomerById(Long id);
-    Customer updateCustomer(Long id, Customer customer);
+    Customer updateCustomer(Integer id, Customer customer);
 }
